Revert "[refactor cluster-task-manage 3/n] Separate stats reporting into its own file (#22359)"
This reverts commit 4ad1fba100752ed166f2a649dbe4831fbf3ae820.
